requirements.txt
# NSE ↔ BSE Arbitrage Trading Bot Dependencies

# Angel One SmartAPI
smartapi-python>=1.3.0

# TOTP for two-factor authentication
pyotp>=2.9.0

# Logging utilities
logzero>=1.7.0

# Terminal dashboard UI
rich>=13.0.0

# Web dashboard (optional)
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# WebSocket support
websocket-client>=1.6.0

# Cryptographic functions
pycryptodome>=3.19.0

